% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDomainInfo.R
\name{addDomainInfo}
\alias{addDomainInfo}
\title{Add domain information to the return (sf) dataframe}
\usage{
addDomainInfo(df, url, token = NULL, call = caller_env())
}
\arguments{
\item{df}{The returned (sf) df from esri2sf/esri2df.}

\item{url}{The url for the Map/Feature server layer/table.}

\item{token}{string for authentication token (if needed).}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
An {sf} dataframe
}
\description{
Uses the esrimeta function to get information about the field
domains in the returned dataframe and joins the domain information to the
dataframe. This is especially important for the codedValue type domains that
are essentially factors and you loose information without replacing the data.
}
