% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriUrl.R
\name{build_esri_community_url}
\alias{build_esri_community_url}
\title{Build a ArcGIS Community API URL}
\usage{
build_esri_community_url(
  url,
  user_id = NULL,
  base_url = "https://www.arcgis.com/sharing/rest/community",
  type = "users",
  sep = "/"
)
}
\description{
Build a ArcGIS Community API URL
}
\keyword{internal}
