% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriUrl.R
\name{build_esri_content_url}
\alias{build_esri_content_url}
\title{Build an ESRI item URL for the content or community API}
\usage{
build_esri_content_url(
  url,
  base_url = "https://www.arcgis.com/sharing/rest/content/",
  type = "items",
  item_id = NULL,
  sep = "/"
)
}
\description{
Build an ESRI item URL for the content or community API
}
\keyword{internal}
