% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriUrl.R
\name{convert_esriUrl}
\alias{convert_esriUrl}
\title{Convert ESRI item URL to feature URL}
\usage{
convert_esriUrl(
  url,
  token = NULL,
  from = NULL,
  to = "feature",
  call = caller_env()
)
}
\arguments{
\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Convert ESRI item URL to feature URL
}
\keyword{internal}
