% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri2sf-package.R
\docType{package}
\name{esri2sf-package}
\alias{esri2sf-package}
\alias{_PACKAGE}
\title{esri2sf: Create Simple Features from ArcGIS Server REST API}
\description{
This package enables you to scrape geographic features directly from ArcGIS servers REST API into R as simple features.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yonghah/esri2sf/}
  \item Report bugs at \url{https://github.com/yonghah/esri2sf/issues}
}

}
\author{
\strong{Maintainer}: Hwang Yongha \email{yongha.hwang@gmail.com}

Authors:
\itemize{
  \item Kevin Cazelles [contributor]
  \item Jacob Peterson [contributor]
  \item Eli Pousson (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [contributor]
}

}
\keyword{internal}
