% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esri_version_ref}
\alias{esri_version_ref}
\title{ArcGIS REST API Version Reference}
\format{
A data frame with 19 rows and 2 variables:
\describe{
\item{\code{version}}{version number}
\item{\code{url}}{url for "What's New" documentation}
}
}
\usage{
esri_version_ref
}
\description{
A reference table with links to information on the ArcGIS REST API version:
\url{https://developers.arcgis.com/rest/services-reference/enterprise/what-s-new.htm}
}
\details{
Additional information on REST API versioning:
\url{https://developers.arcgis.com/rest/services-reference/enterprise/rest-api-versioning.htm}
}
\keyword{datasets}
