% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriIndex.R
\name{esriCatalog}
\alias{esriCatalog}
\title{Get information on folders, services, tables, and layers using the Catalog
service}
\usage{
esriCatalog(
  url,
  format = "json",
  token = NULL,
  option = NULL,
  outSR = NULL,
  ...
)
}
\arguments{
\item{url}{A folder, service, or layer URL that can be used with the ArcGIS
REST API.}

\item{format}{Format to use for request. Supported options include "json", "sitemap", or "geositemap"; "html" and "kmz" are not
currently supported.}

\item{token}{String for authentication token; defaults to \code{NULL}.}

\item{option}{Option parameter, "footprints" is the only option.}

\item{outSR}{Output spatial reference.}

\item{...}{Additional parameters passed to \link[httr2:req_url]{httr2::req_url_query}}
}
\description{
\url{https://developers.arcgis.com/rest/services-reference/enterprise/catalog.htm}
}
