% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esrigeocode.R
\name{esrigeocode}
\alias{esrigeocode}
\title{Use an ArcGIS GeocodeServer to geocode an address or reverse geocode
coordinates}
\usage{
esrigeocode(
  url,
  address = NULL,
  coords = NULL,
  score = 0.95,
  n = 1,
  token = NULL,
  crs = getOption("esri2sf.crs", 4326),
  geometry = TRUE,
  ...
)
}
\arguments{
\item{url}{GeocodeServer service url. Required.}

\item{address}{Single line address passed at the "SingleLine" parameter to
the ArcGIS REST API. Specific format may depend on specific server
configuration, Default: \code{NULL}}

\item{coords}{Numeric vector with longitude, latitude coordinates or a sf
object where the centroid is used as the coordinates. Default: NULL}

\item{score}{Accuracy score, if provided only return results with provided
accuracy score or greater Default: 0.95}

\item{n}{Number of candidates to return, Default: 1}

\item{token}{Token, Default: \code{NULL}}

\item{crs}{Coordinate reference system to return.}

\item{geometry}{If \code{TRUE} (default), return a simple feature object. If
\code{FALSE}, return a data frame.}

\item{...}{Additional parameters passed to \link{esriRequest}}
}
\description{
This function supports two of the services available using an ArcGIS
GeocodeServer:
}
\details{
\itemize{
\item Find Address Candidates:
\url{https://developers.arcgis.com/rest/services-reference/enterprise/find-address-candidates.htm}
\item Reverse Geocode:
\url{https://developers.arcgis.com/rest/services-reference/enterprise/reverse-geocode.htm}
}

Provide an address parameter to use Find Address Candidates or a coords
parameter to use Reverse Geocode.
}
