% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esrihub.R
\name{esrihub}
\alias{esrihub}
\title{Get a feed of available data from an ArcGIS Hub site}
\usage{
esrihub(
  url,
  format = c("dcat-us", "dcat-ap", "rss", "definition"),
  simplifyVector = TRUE
)
}
\arguments{
\item{url}{A url for an ArcGIS Hub site}

\item{format}{Format of data to return, See
\url{https://doc.arcgis.com/en/hub/content/federate-data-with-external-catalogs.htm}
for more information. Default: \code{c("dcat-us", "dcat-ap", "rss", "definition")}.}

\item{simplifyVector}{Passed to \code{\link[httr2:resp_body_raw]{httr2::resp_body_json()}} if format is not
rss. Default: \code{TRUE}}
}
\value{
A list of metadata for the public content on the ArcGIS Hub site.
}
\description{
Get a feed in a DCAT US, DCAT AP, RSS, or OGC Definitions format of the
public data or other content published on an ArcGIS Hub site.
}
