% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriInfo.R
\name{esriInfo}
\alias{esriInfo}
\title{Get info from a service url}
\usage{
esriInfo(url, info = NULL, format = NULL, token = NULL, ...)
}
\arguments{
\item{url}{A folder, service, or layer URL that can be used with the ArcGIS
REST API.}

\item{info}{Info service to use. Options include "info", "item", "metadata",
or "thumbnail". If info = "info" (default) basic server information is
displayed and the body of the response is returned invisibly.}

\item{format}{If \code{info = "metadata"}, options include "fgdc" or "iso19139";
no format is required if \code{info} is "item" or "thumbnail".}

\item{token}{String for authentication token; defaults to \code{NULL}.}

\item{...}{Additional parameters passed to \link[httr2:req_url]{httr2::req_url_query}}
}
\description{
Return server info, service item info, metadata, or thumbnail. The {magick}
package is required to return a thumbnail.
}
\details{
Additional documentation:
\url{https://developers.arcgis.com/rest/services-reference/enterprise/info.htm}
}
