% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriitem.R
\name{esriitem}
\alias{esriitem}
\title{Get ESRI item data or metadata}
\usage{
esriitem(
  url,
  type = "data",
  destfile = tempfile(fileext = "pdf"),
  simplifyVector = TRUE,
  ...
)
}
\arguments{
\item{type}{"data", "info", "metadata", "config" (app URLs only)}

\item{destfile}{Destination file used to download item if data is a PDF file.}

\item{simplifyVector}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}

\item{...}{
  Arguments passed on to \code{\link[httr2:resp_body_raw]{httr2::resp_body_xml}}
  \describe{
    \item{\code{resp}}{A response object.}
    \item{\code{check_type}}{Check that response has expected content type? Set to
\code{FALSE} to suppress the automated check}
  }}
}
\value{
A list, a xml document, or the response object from the request.
}
\description{
\code{\link[=esriitem]{esriitem()}} provides partial support from the ArcGIS Content API.
}
