% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri2sf.R
\name{esrimeta}
\alias{esrimeta}
\title{Retrieve layer metadata}
\usage{
esrimeta(url, token = NULL, fields = FALSE, ..., call = caller_env())
}
\arguments{
\item{url}{url to retrieve metadata for.}

\item{token}{String for authentication token; defaults to \code{NULL}.}

\item{fields}{\code{esrimeta} returns data frame with fields if \code{TRUE}. Default
\code{FALSE}.}

\item{...}{Additional parameters passed to \link[httr2:req_url]{httr2::req_url_query}}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Retrieve layer metadata
}
