% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriRequest.R
\name{esriRequest}
\alias{esriRequest}
\title{Use {httr2} to create a request for the ArcGIS REST API}
\usage{
esriRequest(
  url,
  append = NULL,
  f = NULL,
  format = NULL,
  token = NULL,
  .perform = TRUE,
  .cache = FALSE,
  ...
)
}
\arguments{
\item{url}{A folder, service, or layer URL that can be used with the ArcGIS
REST API.}

\item{append}{String to append to url using \link[httr2:req_url]{httr2::req_url_path_append};
defaults to \code{NULL}.}

\item{f, format}{Return format to use as query parameter with
\link[httr2:req_url]{httr2::req_url_query}; defaults to "json".}

\item{token}{String for authentication token; defaults to \code{NULL}.}

\item{.perform}{If \code{TRUE}, perform the request with \link[httr2:req_perform]{httr2::req_perform} and
return the response. If \code{FALSE}, return the request.}

\item{.cache}{If \code{TRUE}, pass a cache folder path created with \link[rappdirs:user_cache_dir]{rappdirs::user_cache_dir} and \code{esri2sf}
package to the path parameter of \link[httr2:req_cache]{httr2::req_cache}.}

\item{...}{Additional parameters passed to \link[httr2:req_url]{httr2::req_url_query}}
}
\description{
Use {httr2} to create a request for the ArcGIS REST API
}
