% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esrisearch.R
\name{esrisearch}
\alias{esrisearch}
\title{Search for items from ArcGIS Online or an ArcGIS Enterprise Server}
\usage{
esrisearch(
  query = NULL,
  bbox = NULL,
  url = NULL,
  num = 50,
  start = 1,
  category_filter = NULL,
  sort = NULL,
  desc = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{query}{Search terms, Default: \code{NULL}. Required if bbox is \code{NULL}.}

\item{bbox}{A \code{bbox}, \code{sf}, or \code{sfc} to use as a spatial filter for search
results. \code{sf} or \code{sfc} objects are converted to bounding boxes using
\code{\link[sf:st_bbox]{sf::st_bbox()}}.  Default: \code{NULL}. Required if query is \code{NULL}.}

\item{url}{ArcGIS Enterprise request URL (e.g.
"https://machine.domain.com/webadaptor/") or ArcGIS Online organization url
(e.g. "https://org.arcgis.com"). If \code{NULL} (default), url is set to
"https://www.arcgis.com"}

\item{num}{Maximum number of results to return, Default: 50 (must be between
1 and 100).}

\item{start}{Start number for returned results, Default: 1 (returns results
starting from the first result).}

\item{category_filter}{Terms to use in searching for items with matching
categories. Only 3 or less terms currently supported, Default: \code{NULL}}

\item{sort}{Field to use for results sort order. Options include "modified",
"title", "created", "type", "owner", "avgrating", "numratings",
"numcomments", "numviews". Default: \code{NULL}.}

\item{desc}{If \code{TRUE}, return results in descending order. If \code{FALSE}
(default), return results in ascending order. Ignored if sort is \code{NULL}.}
}
\value{
A tibble data.frame with results from the item search.
}
\description{
Use a query or bounding box to search for items on ArcGIS Online or on an
ArcGIS Enterprise Server. See
\url{https://developers.arcgis.com/rest/users-groups-and-items/search.htm} for
more information on the Search endpoint for the Portal Directory API (also
known as the Sharing API). A query or bbox argument must be provided.
}
\examples{
\dontrun{
if (interactive()) {
  esrisearch(query = "park")

  esrisearch(query = c("ocean", "basemap"))

  nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"))

  esrisearch(bbox = nc)
}
}
}
