% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriUrl.R
\name{esriUrl_ServerUrl}
\alias{esriUrl_ServerUrl}
\alias{esriUrl_parseUrl}
\alias{esriUrl_isValid}
\alias{esriUrl_isValidID}
\alias{esriUrl_isValidService}
\title{Get parts of the Map/Feature Server URL}
\usage{
esriUrl_ServerUrl(url)

esriUrl_parseUrl(url)

esriUrl_isValid(url, displayReason = FALSE)

esriUrl_isValidID(url, displayReason = FALSE)

esriUrl_isValidService(url, displayReason = FALSE)
}
\arguments{
\item{url}{The url for a Map/Feature server or for a layer/table in a
Map/Feature Server.}

\item{displayReason}{Should the reason for why a url is not valid be displayed.}
}
\value{
Character string of the request part of the url.
}
\description{
A collection of functions that pull select parts out of a
Map/Feature Server URL. All urls should be a form similar to:
\itemize{
\item \verb{https://<host>/<instance>/rest/services/<folderName>/serviceName>/MapServer/<id>}
\item \verb{http://<host>/<instance>/rest/services/serviceName>/MapServer}
\item \verb{<host>/<instance>/rest/services/<folderName>/serviceName>/FeatureServer}
\item \verb{https://<host>/<instance>/rest/services/serviceName>/FeatureServer/<id>}
And having these rules:
\item The scheme: \verb{https://} or \verb{http://} part is optional
\item The \code{host} part is the domain of the url.
\item The \code{instance}  is the first subpage after the domain in the url.
\item The \verb{/rest/services} is the second and third subpage in the url. These are standard for all ESRI REST Services.
\item The \code{folderName} part is optional and indicates the file structure in the REST Service. It consists of all subpages between \verb{/rest/services/} and the \code{serviceName} part.
\item The \code{serviceName} part is the last subpage betore the \verb{/Mapserver} or \verb{/FeatureServer} in the url.
\item \verb{/Mapserver} or \verb{/FeatureServer} specifies whether the file is a map or feature service.
\item The \code{id} is optional and specifies the layer or table in the map service.
}
}
\section{Functions}{
\itemize{
\item \code{esriUrl_ServerUrl}: Full Map/Feature Server URL

\item \code{esriUrl_parseUrl}: Parse Url into parts.

\item \code{esriUrl_isValid}: Check if url is valid for an ESRI REST Service. General to include potential layer id too.

\item \code{esriUrl_isValidID}: Check if url is valid for an ESRI REST Service Layer or Table

\item \code{esriUrl_isValidService}: Check if url is valid for an ESRI REST Service. No layer ID.
}}

