% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriUrl.R
\name{esriurl_servicetype}
\alias{esriurl_servicetype}
\title{Get a service type for ESRI url}
\usage{
esriurl_servicetype(url)
}
\arguments{
\item{url}{An ESRI url matched to "MapServer", "FeatureServer",
"ImageServer", "GeocodeServer", "GeometryServer", "GPServer",
"GeoDataServer", "WFSServer", or "WFCServer" type.}
}
\description{
Get a service type for ESRI url
}
\keyword{internal}
