% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriuser.R
\name{esriuser}
\alias{esriuser}
\title{Get ArcGIS Online user information}
\usage{
esriuser(url = NULL, user_id = NULL)
}
\arguments{
\item{url}{An ArcGIS Online community URL with a user ID. Optional if user_id is supplied.}

\item{user_id}{An ArcGIS Online user ID. Optional if url is supplied.}
}
\description{
Get ArcGIS Online user information
}
