% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriUrl.R
\name{extract_esri_item_id}
\alias{extract_esri_item_id}
\alias{extract_esri_user_id}
\title{Extract an ESRI item or user id from a url}
\usage{
extract_esri_item_id(
  url,
  pattern = c("(?<=id=)[A-Za-z0-9]+", "(?<=appid=)[A-Za-z0-9]+"),
  collapse = "|"
)

extract_esri_user_id(
  url,
  pattern = c("(?<=user=)[A-Za-z0-9\\\\._\\\\-]+",
    "(?<=owner\%3A\%22)[A-Za-z0-9\\\\._\\\\-]+"),
  collapse = "|"
)
}
\arguments{
\item{url}{A url used to extract an item ID or user ID.}

\item{pattern}{Regular expression pattern passed to \code{\link[stringr:str_extract]{stringr::str_extract()}}.}

\item{collapse}{String used to collapse pattern character vector.}
}
\description{
\code{\link[=extract_esri_item_id]{extract_esri_item_id()}} extract a id or appid value from a URL. If a group
url is supplied a group ID is extracted.

\code{\link[=extract_esri_user_id]{extract_esri_user_id()}} extracts a user id from a profile URL or owner id
from a search URL.
}
