% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fmt_epoch_date}
\alias{fmt_epoch_date}
\alias{fmt_epoch_dates}
\title{Helpers for converting epoch dates to \code{POSIXct} dates}
\usage{
fmt_epoch_date(x, tz = "")

fmt_epoch_dates(x, tz = "")
}
\arguments{
\item{x}{Numeric value corresponding to epoch date.}

\item{tz}{a character string.  The time zone specification to be used
    for the conversion, \emph{if one is required}.  System-specific (see
    \link[base]{time zones}), but \code{""} is the current time zone, and
    \code{"GMT"} is UTC (Universal Time, Coordinated).  Invalid values
    are most commonly treated as UTC, on some platforms with a warning.}
}
\description{
\code{\link[=fmt_epoch_date]{fmt_epoch_date()}} converts a single numeric epoch data value to a \code{POSIXct}
class object. \code{\link[=fmt_epoch_dates]{fmt_epoch_dates()}} converts a numeric vector to \code{POSIXct}
dates.
}
\keyword{helper}
