% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri2sf.R
\name{generateOAuthToken}
\alias{generateOAuthToken}
\title{Generate a OAuth token for Arcgis Online}
\usage{
generateOAuthToken(clientId, clientSecret, expiration = 5000)
}
\arguments{
\item{clientId}{string clientId}

\item{clientSecret}{string clientSecret.}
}
\value{
string token

How to obtain clientId and clientSecret is described here:
https://developers.arcgis.com/documentation/core-concepts/security-and-authentication/accessing-arcgis-online-services/
}
\description{
Generate a OAuth token for Arcgis Online
}
