% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{glue_ansi_sql}
\alias{glue_ansi_sql}
\alias{glue_sql_bbox}
\title{Helpers for creating ANSI SQL queries}
\usage{
glue_ansi_sql(..., .con = DBI::ANSI())

glue_sql_bbox(bbox, coords = c("longitude", "latitude"), crs = 4326)
}
\arguments{
\item{...}{Additional parameters passed to \code{\link[glue:glue_sql]{glue::glue_sql()}}}

\item{.con}{[\code{DBIConnection}]: A DBI connection object obtained from
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{bbox}{A \code{bbox}, \code{sfc}, or \code{sf} object to use as bounding box.}

\item{coords}{Column names with longitude and latitude values.}

\item{crs}{Coordinate reference system used for coordinate values.}
}
\description{
Helpers for creating ANSI SQL queries
}
\examples{

values <- c("a", "b", "c")

glue_ansi_sql("letter", " IN ({values*})")

}
\keyword{helper}
