% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriitem.R
\name{is_esri_content_url}
\alias{is_esri_content_url}
\alias{is_esri_item_url}
\alias{is_esri_app_url}
\title{Is URL an ESRI content url?}
\usage{
is_esri_content_url(x)

is_esri_item_url(x)

is_esri_app_url(x)
}
\description{
Is URL an ESRI content url?
}
