% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDomainInfo.R
\name{addDomainInfo}
\alias{addDomainInfo}
\title{Add domain information to the return (sf) dataframe}
\usage{
addDomainInfo(df, url, token = NULL, call = parent.frame())
}
\arguments{
\item{df}{The returned (sf) df from esri2sf/esri2df.}

\item{url}{The url for the Map/Feature server layer/table.}

\item{token}{string for authentication token (if needed).}

\item{call}{Defaults to \code{\link[=parent.frame]{parent.frame()}}. Passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} to
improve error messages when \code{\link[=esrimeta]{esrimeta()}} is called by another function.}
}
\value{
An {sf} dataframe
}
\description{
Uses the esrimeta function to get information about the field
domains in the returned dataframe and joins the domain information to the
dataframe. This is especially important for the codedValue type domains that
are essentially factors and you loose information without replacing the data.
}
