% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri2rast.R
\name{esri2rast}
\alias{esri2rast}
\title{Import data from ArcGIS ImageServer url using the exportImage API}
\usage{
esri2rast(
  url,
  bbox = NULL,
  token = NULL,
  format = "jpgpng",
  adjustAspectRatio = FALSE,
  ...
)
}
\arguments{
\item{url}{ImageServer url}

\item{bbox}{Bounding box for image to return; defaults to \code{NULL}.}

\item{token}{defaults to \code{NULL}.}

\item{format}{defaults to "jpgpng". Options include "jpgpng", "png", "png8",
"png24", "jpg", "bmp", "gif", "tiff", "png32", "bip", "bsq", and "lerc"}

\item{adjustAspectRatio}{defaults to \code{FALSE}}
}
\value{
SpatRaster object from \code{terra::rast}
}
\description{
See the ArcGIS REST API documentation for more information on the exportImage
API
\url{https://developers.arcgis.com/rest/services-reference/enterprise/export-image.htm}
}
