% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriIndex.R
\name{esriCatalog}
\alias{esriCatalog}
\alias{esricatalog}
\title{Get information on folders, services, tables, and layers using the Catalog
service}
\usage{
esriCatalog(url, f = "json", token = NULL, option = NULL, outSR = NULL, ...)

esricatalog(url, f = "json", token = NULL, option = NULL, outSR = NULL, ...)
}
\arguments{
\item{url}{A folder, service, or layer URL that can be used with the ArcGIS
REST API.}

\item{f}{Format to use for request. Supported options include "json",
"sitemap", or "geositemap"; "html" and "kmz" are not currently supported.}

\item{token}{String for authentication token; defaults to \code{NULL}.}

\item{option}{If \code{option = "footprints"} and the url is for a folder, spatial
footprints of all map, feature, and image services in that folder are
returned as a feature collection}

\item{outSR}{Output spatial reference of the geometry returned in footprints;
only supported when \code{option = "footprints"}.}

\item{...}{Additional parameters passed to \link[httr2:req_url]{httr2::req_url_query}}
}
\description{
The Catalog resource from the ArcGIS REST API represents a catalog of folders
and services published on the host. More information:
\url{https://developers.arcgis.com/rest/services-reference/enterprise/catalog.htm}
}
