% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriInfo.R
\name{esriInfo}
\alias{esriInfo}
\title{Get info from a service url}
\usage{
esriInfo(url, info = NULL, format = NULL, token = NULL, ...)
}
\arguments{
\item{url}{A folder, service, or layer URL that can be used with the ArcGIS
REST API.}

\item{info}{Options include "item", "metadata", or "thumbnail".}

\item{format}{If \code{info = "metadata"}, options include "fgdc" or "iso19139";
"json" format is used for \code{info = "item"} and no format is required for
\code{info = "thumbnail"}}

\item{token}{String for authentication token; defaults to \code{NULL}.}

\item{...}{Additional parameters passed to \link[httr2:req_url]{httr2::req_url_query}}
}
\description{
Return service item info, metadata, or thumbnail. The {magick} package is
required to return a thumbnail.
}
\details{
Additional documentation:
\url{https://developers.arcgis.com/rest/services-reference/enterprise/info.htm}
}
