% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriLayers.R
\name{esriLayers}
\alias{esriLayers}
\title{Get All Layers and Tables for the whole server}
\usage{
esriLayers(url)
}
\arguments{
\item{url}{The url for the Map/Feature server. If given a url specifying a
layer or table ID it will truncate it.}
}
\value{
A list from the JSON return.
}
\description{
Retrieve JSON for all layers and tables as specified in
\url{https://si-pweb-vecmap.si.edu/vectormap/sdk/rest/index.html#/All_Layers_and_Tables/02ss0000005t000000/}.
Performs a POST request towards the server url in one of the following the
forms:
\itemize{
\item \verb{https://<host>/<instance>/rest/services/<folderName>/serviceName>/MapServer/layers}
\item \verb{https://<host>/<instance>/rest/services/serviceName>/MapServer/layers}
\item \verb{https://<host>/<instance>/rest/services/<folderName>/serviceName>/FeatureServer/layers}
\item \verb{https://<host>/<instance>/rest/services/serviceName>/FeatureServer/layers}
}
}
