% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri2sf.R
\name{esrimeta}
\alias{esrimeta}
\title{Retrieve layer metadata}
\usage{
esrimeta(url, token = NULL, fields = FALSE, ..., call = parent.frame())
}
\arguments{
\item{url}{url to retrieve metadata for.}

\item{token}{String for authentication token; defaults to \code{NULL}.}

\item{fields}{\code{esrimeta} returns data frame with fields if \code{TRUE}. Default
\code{FALSE}.}

\item{...}{Additional parameters passed to \link[httr2:req_url]{httr2::req_url_query}}

\item{call}{Defaults to \code{\link[=parent.frame]{parent.frame()}}. Passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} to
improve error messages when \code{\link[=esrimeta]{esrimeta()}} is called by another function.}
}
\description{
Retrieve layer metadata
}
