% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esriRequest.R
\name{esriRequest}
\alias{esriRequest}
\title{Use {httr2} to create a request for the ArcGIS REST API}
\usage{
esriRequest(
  url,
  append = NULL,
  format = "json",
  token = NULL,
  perform = TRUE,
  ...
)
}
\arguments{
\item{url}{A folder, service, or layer URL that can be used with the ArcGIS
REST API.}

\item{token}{String for authentication token; defaults to \code{NULL}.}

\item{...}{Additional parameters passed to \link[httr2:req_url]{httr2::req_url_query}}
}
\description{
Use {httr2} to create a request for the ArcGIS REST API
}
