% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{generateToken}
\alias{generateToken}
\alias{generateOAuthToken}
\title{Create authorization tokens}
\usage{
generateToken(
  server,
  uid,
  pwd = "",
  type = c("tokens", "admin"),
  expiration = 5000
)

generateOAuthToken(clientId, clientSecret, expiration = 5000)
}
\arguments{
\item{server}{The ArcGIS REST Server you want to connect to: \verb{https://\\<host\\>:\\<port\\>}}

\item{uid}{The user id of the account used to create the token connection to
the server}

\item{pwd}{The password of the account used to create the token connection to
the server. If left blank, you will be prompted for the password.}

\item{type}{Either 'tokens' or 'admin'. Specify the endpoint you use to
create the token. Defaults to 'tokens'.}

\item{expiration}{Set an expiration limit on the token in minutes. Max
expiration date may be controlled by the server.}

\item{clientId}{Client ID}

\item{clientSecret}{Client Secret}
}
\value{
Character string with the token
}
\description{
Generate tokens for accessing credentialed ArcGIS REST Servers.

\code{generateToken()} can create a token from the public token endpoint
\verb{https://<host>:\\<port\\>/\\<site\\>/tokens/generateToken} or the admin endpoint
\verb{https://\\<host\\>:\\<port\\>/\\<site\\>/admin/generateToken} for ArcGIS REST Servers. See
\verb{https://developers.arcgis.com/rest/services-reference/enterprise/generate-token.htm} or
\verb{https://developers.arcgis.com/rest/services-reference/enterprise/generate-admin-token.htm}
respectively for more information.

\code{generateOAuthToken()} can create an OAuth token for ArcGIS Online Services.
See \verb{https://developers.arcgis.com/documentation/core-concepts/security-and-authentication/accessing-arcgis-online-services/}
or \verb{https://developers.arcgis.com/documentation/mapping-apis-and-services/security/oauth-2.0/}
for more information.
}
\section{Functions}{
\itemize{
\item \code{generateToken}: Create ArcGIS REST Service Token

\item \code{generateOAuthToken}: Create ArcGIS OAuth Token
}}

