% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlParts.R
\name{urlParts_urlServer}
\alias{urlParts_urlServer}
\alias{urlParts_host}
\alias{urlParts_instance}
\alias{urlParts_folder}
\alias{urlParts_service}
\alias{urlParts_layerID}
\title{Get parts of the Map/Feature Server URL}
\usage{
urlParts_urlServer(url)

urlParts_host(url)

urlParts_instance(url)

urlParts_folder(url)

urlParts_service(url)

urlParts_layerID(url)
}
\arguments{
\item{url}{The url for a Map/Feature server or for a layer/table in a
Map/Feature Server.}
}
\value{
Character string of the request part of the url.
}
\description{
A collection of functions that pull select parts out of a
Map/Feature Server URL.
}
\details{
The url should be formatted similar to these examples:
\itemize{
\item \verb{https://<host>/<instance>/rest/services/<folderName>/serviceName>/MapServer/<id>}
\item \verb{https://<host>/<instance>/rest/services/serviceName>/MapServer}
\item \verb{https://<host>/<instance>/rest/services/<folderName>/serviceName>/FeatureServer}
\item \verb{https://<host>/<instance>/rest/services/serviceName>/FeatureServer/<id>}
}
}
\section{Functions}{
\itemize{
\item \code{urlParts_urlServer}: Full Map/Feature Server URL

\item \code{urlParts_host}: URL host

\item \code{urlParts_instance}: URL instance

\item \code{urlParts_folder}: URL folder(s)

\item \code{urlParts_service}: URL service

\item \code{urlParts_layerID}: Layer/Table ID
}}

